Moving amp-ad-template to amp-ad-custom.
